requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-telegram-bot>=20.0,<22.0
python-multipart>=0.0.9,<0.1
jinja2>=3.1,<4
pydantic-settings>=2.0,<3
sse-starlette>=1.5,<4
httpx>=0.26,<1.0
passlib[bcrypt]>=1.7.4,<2.0

# Development tools
ruff>=0.6,<1.0

# Testing tools
locust>=2.0,<3.0
